tidy up after merge
