+++ migration_steps/integration/reindex_ids/app/app.py
@@ -5,7 +5,7 @@ import logging
 import time
 import click
 from sqlalchemy import create_engine
-from existing_data.match_existing_data_fresh import match_existing_data, fks
+from existing_data.match_existing_data import match_existing_data
 from reindex.move_by_table import move_all_tables, create_schema
 from reindex.reindex_foreign_keys import update_fks
 from reindex.reindex_primary_keys import update_pks
